move state file controller concern stuff into controller and delete concern
